[No ticket] Cambios para dejar de usar `navigate(0)` en frontend (#106)
