+++ src/db/seeds.sql
@@ -35,7 +35,7 @@ SELECT setval('teachhub.subjects_id_seq', (select MAX(id) from teachhub.subjects
 --
 
 INSERT INTO teachhub.roles (id, name, parent_role_id, permissions, active, is_teacher) VALUES (1, 'Usuario', NULL, 'viewHome', false, false);
-INSERT INTO teachhub.roles (id, name, parent_role_id, permissions, active, is_teacher) VALUES (2, 'Titular', NULL, 'setOrganization, createRepository, editAssignment, createAssignment, inviteUser, editSubject, assignReviewer, viewGroups', false, true);
+INSERT INTO teachhub.roles (id, name, parent_role_id, permissions, active, is_teacher) VALUES (2, 'Titular', NULL, 'setOrganization, createRepository, editAssignment, createAssignment, inviteUser, editSubject, assignReviewer, viewGroups, viewAllSubmissions', false, true);
 INSERT INTO teachhub.roles (id, name, parent_role_id, permissions, active, is_teacher) VALUES (3, 'JTP', NULL, '', false, true);
 INSERT INTO teachhub.roles (id, name, parent_role_id, permissions, active, is_teacher) VALUES (4, 'Ayudante', NULL, '', false, true);
 INSERT INTO teachhub.roles (id, name, parent_role_id, permissions, active, is_teacher) VALUES (5, 'Alumno', NULL, 'manageOwnGroups, submitAssignment, viewSubmission', false, false);
